+++ app/layout.tsx
@@ -4,6 +4,7 @@ import './globals.css'
 import Providers from '../src/app/providers'
 import { Layout } from '@/shared/ui'
 import Script from 'next/script'
+import { UserStateRouter } from '@/features/routing'
 
 export const viewport: Viewport = {
   width: 'device-width',
@@ -45,6 +46,7 @@ export default function RootLayout({
       </head>
       <body>
         <Providers>
+          <UserStateRouter />
           <Layout>{children}</Layout>
         </Providers>
       </body>
